Extract OutcomeHandler to the package level
